Add offline notification. Closes #9.
Meteor.status() information added to the bottom of the screen when the server is offline or unavailable. Includes a retry counter indicating how many attempts have been made, how many seconds until the next update, and a button to attempt to reconnect immediately.

Future improvements may include a live counter actually counting down to when the reconnect will happen.