Add sensorInfo control to core.
Having issues with card rendering when updating the simulator. We're either going to have to figure out the problem and fix it or split the Simulator and Station models into two pieces, with a "SimulatorLayout" and "StationLayout" counterpart that provides the data backing to the interface, e.g. layouts and themes. In this way, the theme and layout won't be redrawn when we change anything about the station or the simulator, thus avoiding the "white flash" issue. Refs #2, refs #11, refs #10.