Add sensor grid to Core. Refs #10.
No controls (yet) but the sensor grid mirrors now between core and station. A few tweaks to velocity handling to accommodate for different scaling.

As part of this process, I realized that the SensorGrid is more of a UI component that ought to be droppable anywhere. More thought should go into how this ought to be implemented (while similar, a core has admin rights while a station just watches). Even so, we need to abstract the sensor functionality somehow. Refs #14.