Add shadow contacts to core sensor grid. Refs #10.
Flight directors can now tap the core grid to view the shadow contacts. Runs its own animation loop, the animation loop is running constantly, regardless of whether or not we're viewing shadow contacts. If we want/need to optimize this later, we could track the time delta between starting movement and showing the animation loop, however I'm opting for functionality before performance in this case.