Implement core station utility to manipulate stations. Closes #20.
There are a lot of methods to be implemented on this one, but it presently enables a flight director to flash or trigger a station reelection on either a participant or a station. If a station is loaded in multiple locations, all stations flash (hence the "participant" feature to target an individual if desired).