Added Three.js. Included a couple of packages for content right off the bat.
To make it work, I did some changes to the three.js file in the three.js package, which is downloaded by Meteorite. 1st updated the file from threejs.com. 2nd added 'window.THREE = THREE' near the top of the file.

Added the model into the thruster screen and made it move when the thrusters are fired.