Rework of Actor API to handle multiple contacts better, remove contacts when at 120% of radius, and to otherwise perform well.
This actor will be ported to GSAP and we will discontinue Tween.js for Flint. GSAP supports a number of advanced tweening features that are necessary for sensors to have all the functionality we want/need, including in-flight changes to destination Tweens. We may still end up writing our own physics engine, but hopefully GSAP can handle what we need from it in the meantime.